+++ src/components/modals/SettingsModal.js
@@ -1,4 +1,4 @@
-import { exitToMenu, restartGame } from '../utils/utils';
+import { exitToMenu, resetGame } from '../utils/utils';
 import { click } from '../utils/assetPaths';
 
 const SettingsModal = ({
@@ -63,9 +63,9 @@ const SettingsModal = ({
                 <button
                   type='button'
                   className='btn btn-dark btn-width mb-3'
-                  onClick={restartGame}
+                  onClick={resetGame}
                 >
-                  Restart Game
+                  Reset Game
                 </button>
 
                 <button