+++ CHANGELOG.md
@@ -1,10 +1,19 @@
+# 🚀 Release v0.12 (2024-11-25)
+
+New Feature: Levels
+
+- Introduce levels to the game. Each level has unique artwork and mechanics to increase variety and player experience.
+- Level 1: Royal City
+- Level 2: Lava Plains
+
 # 🚀 Release v0.11 (2024-08-11)
 
 UX improvements.
-* Add settings menu - [8d59ebb](https://github.com/ruichen199801/wizard-duel/commit/8d59ebbd1407787b92c407cff28676d5c60e8903)
-* Add background music and music & sound controls - [3fc9ca6](https://github.com/ruichen199801/wizard-duel/commit/3fc9ca6690feb9a7dedbccc8b334d88bb1187d1a)
-* Add a matchup screen when game starts - [183ae6c](https://github.com/ruichen199801/wizard-duel/commit/183ae6cc6ce18fdc30266b960331c1d2483953df)
-* Add a list of active effects that can be viewed by hovering over each player's avatar - [62e82ad](https://github.com/ruichen199801/wizard-duel/commit/62e82adfe025bac00108fb135efecd53a5ea89fb)
+
+- Add settings menu - [8d59ebb](https://github.com/ruichen199801/wizard-duel/commit/8d59ebbd1407787b92c407cff28676d5c60e8903)
+- Add background music and music & sound controls - [3fc9ca6](https://github.com/ruichen199801/wizard-duel/commit/3fc9ca6690feb9a7dedbccc8b334d88bb1187d1a)
+- Add a matchup screen when game starts - [183ae6c](https://github.com/ruichen199801/wizard-duel/commit/183ae6cc6ce18fdc30266b960331c1d2483953df)
+- Add a list of active effects that can be viewed by hovering over each player's avatar - [62e82ad](https://github.com/ruichen199801/wizard-duel/commit/62e82adfe025bac00108fb135efecd53a5ea89fb)
 
 # 🚀 Release v0.1 (2024-07-24)